[compiler] Extract reusable logic for control dominators
The next PR needs to check if a block is controlled by a value derived from a ref.
